mm: Change prototype of mm_sbrk()
Make its parameter a size_t instead of an unsigned long.

+++ include/kos/mm.h
@@ -19,6 +19,8 @@
 #include <kos/cdefs.h>
 __BEGIN_DECLS
 
+#include <stddef.h>
+
 /** \brief   Initialize the memory management system.
     \ingroup mm
 
@@ -33,7 +35,7 @@ int mm_init(void);
     \return                 A pointer to the memory.
     \note                   This function will panic if no memory is available.
 */
-void *mm_sbrk(unsigned long increment);
+void *mm_sbrk(size_t increment);
 
 __END_DECLS
 #endif /* __KOS_MM_H */
