+++ kernel/libc/newlib/newlib_sbrk.c
@@ -10,5 +10,5 @@
 
 char * _sbrk_r(struct _reent * reent, size_t incr) {
     (void)reent;
-    return (char *)mm_sbrk((unsigned long)incr);
+    return (char *)mm_sbrk(incr);
 }
