+++ kernel/mm/mm.c
@@ -38,7 +38,7 @@ int mm_init(void) {
 }
 
 /* Simple sbrk function */
-void *mm_sbrk(unsigned long increment) {
+void *mm_sbrk(size_t increment) {
     uintptr_t base = sbrk_base;
 
     irq_disable_scoped();
@@ -49,7 +49,7 @@ void *mm_sbrk(unsigned long increment) {
 
     if(sbrk_base >= (_arch_mem_top - THD_KERNEL_STACK_SIZE)) {
         dbglog(DBG_CRITICAL, "Out of memory. Requested sbrk_base %" PRIxPTR \
-            ", was %" PRIxPTR ", diff %lu\n", sbrk_base, base, increment);
+            ", was %" PRIxPTR ", diff %zu\n", sbrk_base, base, increment);
         sbrk_base = base;  /* Restore old value and mark failed */
         errno = ENOMEM;
         return (void *)-1;
